absolutely stupid and meaningless refactoring to calm down Bohdan
