+++ server/rules.go
@@ -52,71 +52,6 @@ func (server *HTTPServer) getRule(writer http.ResponseWriter, request *http.Requ
 	}
 }
 
-// likeRule likes the rule for current user
-func (server *HTTPServer) likeRule(writer http.ResponseWriter, request *http.Request) {
-	server.voteOnRule(writer, request, types.UserVoteLike)
-}
-
-// dislikeRule dislikes the rule for current user
-func (server *HTTPServer) dislikeRule(writer http.ResponseWriter, request *http.Request) {
-	server.voteOnRule(writer, request, types.UserVoteDislike)
-}
-
-// resetVoteOnRule resets vote for the rule for current user
-func (server *HTTPServer) resetVoteOnRule(writer http.ResponseWriter, request *http.Request) {
-	server.voteOnRule(writer, request, types.UserVoteNone)
-}
-
-func (server *HTTPServer) voteOnRule(writer http.ResponseWriter, request *http.Request, userVote types.UserVote) {
-	clusterID, ruleID, userID, err := server.readClusterRuleUserParams(writer, request)
-	if err != nil {
-		// everything has been handled already
-		return
-	}
-
-	err = server.checkUserClusterPermissions(writer, request, clusterID)
-	if err != nil {
-		// everything has been handled already
-		return
-	}
-
-	err = server.Storage.VoteOnRule(clusterID, ruleID, userID, userVote)
-	if err != nil {
-		handleServerError(writer, err)
-		return
-	}
-
-	err = responses.SendOK(writer, responses.BuildOkResponse())
-	if err != nil {
-		log.Error().Err(err).Msg(responseDataError)
-	}
-}
-
-func (server *HTTPServer) getVoteOnRule(writer http.ResponseWriter, request *http.Request) {
-	clusterID, ruleID, userID, err := server.readClusterRuleUserParams(writer, request)
-	if err != nil {
-		// everything has been handled already
-		return
-	}
-
-	err = server.checkUserClusterPermissions(writer, request, clusterID)
-	if err != nil {
-		// everything has been handled already
-		return
-	}
-
-	userFeedbackOnRule, err := server.Storage.GetUserFeedbackOnRule(clusterID, ruleID, userID)
-	if err != nil {
-		handleServerError(writer, err)
-		return
-	}
-
-	err = responses.SendOK(writer, responses.BuildOkResponseWithData("vote", userFeedbackOnRule.UserVote))
-	if err != nil {
-		log.Error().Err(err).Msg(responseDataError)
-	}
-}
-
 // disableRuleForCluster disables a rule for specified cluster, excluding it from reports
 func (server *HTTPServer) disableRuleForCluster(writer http.ResponseWriter, request *http.Request) {
 	server.toggleRuleForCluster(writer, request, storage.RuleToggleDisable)
